Large refactor in prep for mainnet deployment
Several changes to split mainnet from devnet constants. Additionally, change message PDA to be based on mailbox rather than receiver to allow for additional ways to generate a mailbox in the future.

Remove seed constants and instead use anchor automatic seed logic now that it is available

Update docs for new API and add changelog
