+++ programs/messaging/Cargo.toml
@@ -13,6 +13,7 @@ no-entrypoint = []
 no-idl = []
 cpi = ["no-entrypoint"]
 default = []
+mainnet = []
 
 [dependencies]
 anchor-lang = {version = "0.22.0", features = ["init-if-needed"]}
