+++ programs/messaging/src/treasury.rs
@@ -1,8 +1,6 @@
 use anchor_lang::prelude::*;
 
-#[cfg(feature = "devnet")]
-#[constant]
-pub const TREASURY_ADDRESS: Pubkey = solana_program::pubkey!("G2GGDc89qpuk21WgRUVPDY517uc6qR5yT4KX7AakyVR1");
-#[cfg(not(feature = "devnet"))]
-#[constant]
+#[cfg(feature = "mainnet")]
+pub const TREASURY_ADDRESS: Pubkey = solana_program::pubkey!("5MNBoBJDHHG4pB6qtWgYPzGEncoYTLAaANovvoaxu28p");
+#[cfg(not(feature = "mainnet"))]
 pub const TREASURY_ADDRESS: Pubkey = solana_program::pubkey!("G2GGDc89qpuk21WgRUVPDY517uc6qR5yT4KX7AakyVR1");
